Project Update

+++ team_AK/src/main/java/internet_store/core/request/product/product_items/AddProductQuantityRequest.java
@@ -2,13 +2,11 @@ package internet_store.core.request.product.product_items;
 
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 public class AddProductQuantityRequest {
     @Getter
-    private final BigDecimal quantity;
+    private final Long quantity;
 
-    public AddProductQuantityRequest(BigDecimal quantity) {
+    public AddProductQuantityRequest(Long quantity) {
         this.quantity = quantity;
     }
 }
